Added helper functions for slackbot
